+++ src/App.js
@@ -2,12 +2,12 @@ import MyReservePage from "./pages/MyPage/MyReservePage";
 import DetailReservePage from "./pages/MyPage/DetailReservePage";
 import "./styles/Font.css";
 import Header from "./components/common/Header";
-import ReservationDetail from "./pages/photographer/ReservationDetail";
-import Reservation from "./pages/photographer/Reservation";
-import MyPage from "./pages/photographer/MyPage";
-import Setting from "./pages/photographer/Setting";
-import Custom from "./pages/photographer/Custom";
-import Review from "./pages/photographer/Review";
+import ReservationDetail from "./pages/Photographer/ReservationDetail";
+import Reservation from "./pages/Photographer/Reservation";
+import MyPage from "./pages/Photographer/MyPage";
+import Setting from "./pages/Photographer/Setting";
+import Custom from "./pages/Photographer/Custom";
+import Review from "./pages/Photographer/Review";
 import SearchPage from "./pages/serach/SearchPage";
 import { Routes, Route } from "react-router-dom";
 import Footer from "./components/common/Footer";
