+++ src/components/Photographer/Setting/Modal.js
